+++ extensions/safari/src/utils.ts
@@ -3,7 +3,7 @@ import Fuse, { FuseOptionKey } from "fuse.js";
 import _ from "lodash";
 import osascript from "osascript-tag";
 import { URL } from "url";
-import { langAdaptor, PinyinHandler } from "./lang-adaptor";
+import { langAdaptor } from "./lang-adaptor";
 import { HistoryItem, LooseTab } from "./types";
 import { runAppleScript } from "@raycast/utils";
 
@@ -70,7 +70,13 @@ export const getTitle = (tab: LooseTab) => _.truncate(tab.title, { length: 75 })
 export const plural = (count: number, string: string) => `${count} ${string}${count > 1 ? "s" : ""}`;
 
 function installLangHandlers() {
-  langAdaptor.registerLang(PinyinHandler.name, new PinyinHandler());
+  const enablePinyin = getPreferenceValues().enablePinyin;
+  if (enablePinyin) {
+    import("./lang-adaptor/pinyin").then((pinyinModule) => {
+      const pinyinHandler = new pinyinModule.PinyinHandler();
+      langAdaptor.registerLang(pinyinHandler.name, pinyinHandler);
+    });
+  }
 }
 
 export const search = function (collection: LooseTab[], keys: Array<FuseOptionKey<object>>, searchText: string) {